git-svn-id: http://svn.wp-plugins.org/wordpress-flickr-manager/trunk@85024 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ js/media-panel.php
@@ -265,9 +265,9 @@ var prepareImages = function() {
 				jQuery("body").append(wrapper);
 				wrapper.css(listCSS);
 				var minWidth = 135;
-				if(parseInt(wrapper.css('left').match(/\d+/)) + minWidth > jQuery("body").width()) {
+				if(parseFloat(wrapper.css('left').match(/\d+/)) + minWidth > jQuery("body").width()) {
 					wrapper.css('width', minWidth);
-					wrapper.css('left', parseInt(wrapper.css('left').match(/\d+/)) - minWidth);
+					wrapper.css('left', parseFloat(wrapper.css('left').match(/\d+/)) - minWidth);
 				}
 				
 				jQuery('div.sizePicker>a.closeLink').click(function() {
